Move data models into dedicated files
